Globalisation

+++ app/src/main/java/com/umangSRTC/thesohankathait/classes/Fragment/Schools.java
@@ -11,7 +11,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.AdapterView;
-import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ListView;
@@ -90,7 +89,7 @@ public class Schools extends Fragment {
         schoolsListView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
             @Override
             public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-                if(Admin.CheckAdmin(FirebaseAuth.getInstance().getCurrentUser().getEmail()))
+                if(Admin.CheckAdmin(FirebaseAuth.getInstance().getCurrentUser().getEmail()) && !Initialisation.schoolArrayList.get(position).equals("Notices"))
                     deleteWarning(Initialisation.schoolArrayList.get(position),position);
                 return true;
             }
@@ -130,7 +129,7 @@ public class Schools extends Fragment {
 
     private void deleteSchoolFromFirebase(final String school) {
 
-        FirebaseDatabase.getInstance().getReference("Schools").addValueEventListener(new ValueEventListener() {
+        FirebaseDatabase.getInstance().getReference(Initialisation.selectedCollege+"/Schools").addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
                 for(DataSnapshot finalDataSnapshot:dataSnapshot.getChildren()){
@@ -156,7 +155,7 @@ public class Schools extends Fragment {
 
     private void deleteFromValue(final String school, String category) {
 
-        FirebaseDatabase.getInstance().getReference(category).child(school).removeValue();
+        FirebaseDatabase.getInstance().getReference(Initialisation.selectedCollege+"/"+category).child(school).removeValue();
 
     }
 
@@ -196,8 +195,8 @@ public class Schools extends Fragment {
     }
 
     private boolean addSchoolIntoFirebase(String schoolName) {
-        if(Admin.isSchoolCorrect(schoolName)){
-            FirebaseDatabase.getInstance().getReference("Schools").push().setValue(schoolName);
+        if(Admin.isCorrect(schoolName)){
+            FirebaseDatabase.getInstance().getReference(Initialisation.selectedCollege+"/Schools").push().setValue(schoolName);
             return true;
         }
         else{
